ERT-1059 Moved Legend widget
