+++ devel/python/python/ert_gui/ertwidgets/legend.py
